Refactor communication strategy and add vehicle configuration
Removed some redundant code and restructured the communication strategy in the custom adapter. Also included configurations for vehicles, which allows us to initialize strategies and receive specific communication from different vehicle strategies. To better handle strategy creation, StrategyCreator interface and classes were introduced, and a Vehicle configuration was added for personalized vehicle settings.
