+++ openTCS-CustomAdapter/src/main/java/org/opentcs/customadapter/CustomCommunicationAdapterFactory.java
@@ -2,20 +2,11 @@ package org.opentcs.customadapter;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
+import com.google.inject.assistedinject.AssistedInject;
 import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.ScheduledExecutorService;
 import java.util.logging.Logger;
-import javax.swing.BoxLayout;
-import javax.swing.JComboBox;
-import javax.swing.JLabel;
-import javax.swing.JOptionPane;
-import javax.swing.JPanel;
-import javax.swing.JTextField;
 import org.opentcs.components.kernel.services.PlantModelService;
-import org.opentcs.customizations.kernel.KernelExecutor;
 import org.opentcs.data.model.Vehicle;
 import org.opentcs.drivers.vehicle.VehicleCommAdapter;
 import org.opentcs.drivers.vehicle.VehicleCommAdapterDescription;
@@ -44,6 +35,7 @@ public class CustomCommunicationAdapterFactory
    * @param plantModelService The service used for interacting with plant models.
    */
   @Inject
+  @AssistedInject
   public CustomCommunicationAdapterFactory(
       CustomAdapterComponentsFactory adapterFactory,
       VehicleConfigurationProvider configProvider,
@@ -97,167 +89,5 @@ public class CustomCommunicationAdapterFactory
   Vehicle vehicle) {
     LOG.info("Creating adapter for vehicle: " + vehicle.getName());
     return adapterFactory.createCustomCommAdapter(vehicle);
-  }
-
-  private static class CustomVehicleCommAdapterDescription
-      extends
-        VehicleCommAdapterDescription {
-    CustomVehicleCommAdapterDescription() {
-    }
-
-    @Override
-    public String getDescription() {
-      return "Custom Vehicle Communication Adapter";
-    }
-
-    @Override
-    public boolean isSimVehicleCommAdapter() {
-      return false;
-    }
-  }
-}
-
-@Singleton
-class CommunicationStrategy
-    implements
-      CustomAdapterComponentsFactory {
-  private static final Logger LOG = Logger.getLogger(CommunicationStrategy.class.getName());
-
-  private final Map<String, StrategyCreator> strategies = new HashMap<>();
-  private final VehicleConfigurationProvider configProvider;
-  private final ScheduledExecutorService executor;
-  private final PlantModelService plantModelService;
-
-  @Inject
-  CommunicationStrategy(
-      @KernelExecutor
-      ScheduledExecutorService executor,
-      VehicleConfigurationProvider configProvider,
-      PlantModelService plantModelService
-  ) {
-    this.executor = executor;
-    this.configProvider = configProvider;
-    this.plantModelService = plantModelService;
-    initializeStrategies();
-  }
-
-  @SuppressWarnings("checkstyle:TodoComment")
-  private void initializeStrategies() {
-    strategies.put("ModbusTCP", new ModbusTCPStrategy());
-    // TODO: Add other strategies here
-  }
-
-  @Override
-  public CustomVehicleCommAdapter createCustomCommAdapter(Vehicle vehicle) {
-    VehicleConfiguration config = configProvider.getConfiguration(vehicle.getName());
-    if (config == null) {
-      config = createConfigWithUserInput(vehicle);
-      configProvider.setConfiguration(vehicle.getName(), config);
-    }
-
-    String strategyKey = config.currentStrategy();
-    StrategyCreator creator = strategies.get(strategyKey);
-    if (creator == null) {
-      LOG.warning("Unknown strategy: " + strategyKey + ". Using default ModbusTCP strategy.");
-      creator = strategies.get("ModbusTCP");
-    }
-
-    return creator.createAdapter(vehicle, config, executor, plantModelService);
-  }
-
-  private VehicleConfiguration createConfigWithUserInput(Vehicle vehicle) {
-    String defaultHost = "localhost";
-    int defaultPort = 502;
-    String defaultStrategy = "ModbusTCP";
-
-    JTextField hostField = new JTextField(defaultHost, 10);
-    JTextField portField = new JTextField(String.valueOf(defaultPort), 10);
-    JComboBox<String> strategyComboBox = new JComboBox<>(
-        strategies.keySet().toArray(new String[0])
-    );
-    strategyComboBox.setSelectedItem(defaultStrategy);
-
-    JPanel panel = new JPanel();
-    panel.setLayout(new BoxLayout(panel, BoxLayout.Y_AXIS));
-    panel.add(new JLabel("Host:"));
-    panel.add(hostField);
-    panel.add(new JLabel("Port:"));
-    panel.add(portField);
-    panel.add(new JLabel("Communication Strategy:"));
-    panel.add(strategyComboBox);
-
-    int result = JOptionPane.showConfirmDialog(
-        null, panel,
-        "Enter Configuration for " + vehicle.getName(), JOptionPane.OK_CANCEL_OPTION
-    );
-
-    String host;
-    int port;
-    String strategy;
-
-    if (result == JOptionPane.OK_OPTION) {
-      host = hostField.getText().isEmpty() ? defaultHost : hostField.getText();
-      try {
-        port = Integer.parseInt(portField.getText());
-        if (port < 0 || port > 65535) {
-          LOG.warning("Invalid port number. Using default port " + defaultPort);
-          port = defaultPort;
-        }
-      }
-      catch (NumberFormatException e) {
-        LOG.warning("Invalid port number. Using default port " + defaultPort);
-        port = defaultPort;
-      }
-      strategy = (String) strategyComboBox.getSelectedItem();
-      if (strategy == null || strategy.isEmpty()) {
-        strategy = defaultStrategy;
-      }
-    }
-    else {
-      // Use default values if user cancels
-      host = defaultHost;
-      port = defaultPort;
-      strategy = defaultStrategy;
-    }
-
-    return new VehicleConfiguration(strategy, host, port);
-  }
-}
-
-interface StrategyCreator {
-  CustomVehicleCommAdapter createAdapter(
-      Vehicle vehicle,
-      VehicleConfiguration config,
-      ScheduledExecutorService executor,
-      PlantModelService plantModelService
-  );
-}
-
-class ModbusTCPStrategy
-    implements
-      StrategyCreator {
-  private static final String DEFAULT_RECHARGE_OPERATION = "RECHARGE";
-  private static final int DEFAULT_COMMANDS_CAPACITY = 1000;
-
-  ModbusTCPStrategy() {
-  }
-
-  @Override
-  public CustomVehicleCommAdapter createAdapter(
-      Vehicle vehicle,
-      VehicleConfiguration config,
-      ScheduledExecutorService executor,
-      PlantModelService plantModelService
-  ) {
-    return new ModbusTCPVehicleCommAdapter(
-        new CustomProcessModel(vehicle),
-        DEFAULT_RECHARGE_OPERATION,
-        DEFAULT_COMMANDS_CAPACITY,
-        executor,
-        vehicle,
-        config.host(),
-        config.port(),
-        plantModelService
-    );
   }
 }
